% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{beta.mm}
\alias{beta.mm}
\title{Compute the least-squares slope estimate for simple linear regression.}
\usage{
beta.mm(x, y)
}
\arguments{
\item{x}{A numeric vector containing entries for the independent variable.}

\item{y}{A numeric vector containing entries for the dependent variable, must be the same length as x.}
}
\value{
The least-squares slope estimate (i.e. the estimate of b) for the model y = a + b*x + disturbance.
}
\description{
Given two numeric vectors, computes a least-squares slope estimate for the model y = a + b*x + disturbance. lm() also returns this number, but beta.mm() is faster, which is advantageous for simulating bootstrapping or permutation testing.
}
\examples{
beta.mm(rnorm(10), rnorm(10))
}
\keyword{least-squares,}
\keyword{linear}
\keyword{method}
\keyword{moments}
\keyword{of}
\keyword{regression,}
\keyword{simple}
