% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{boot.dist.1d}
\alias{boot.dist.1d}
\title{Generate a bootstrap distribution for a statistic computed on a vector.}
\usage{
boot.dist.1d(x, B, FUN = mean, ...)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{B}{The number of bootstrap samples to draw.}

\item{FUN}{A function that computes the statistic of interest. FUN must have a vector of data as its first argument and must return a numeric vector of length 1 (i.e. a scalar).}

\item{...}{Additional arguments passed to FUN.}
}
\value{
A numeric vector containing the bootstrap distribution of the statistic specified by FUN.
}
\description{
Given data x, generates a bootstrap distribution with B bootstrap samples, computing a statistic speficied by FUN each time.
}
\examples{
boot.dist.1d(rnorm(50, 0, 1), 1000, median)
}
\keyword{Bootstrap}
\keyword{Bootstrapping,}
\keyword{distribution}
