% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{boot.samp}
\alias{boot.samp}
\title{Draw a bootstrap sample from a matrix or vector}
\usage{
boot.samp(x)
}
\arguments{
\item{x}{A matrix or vector.}
}
\value{
A matrix containing a bootstrap sample of the original matrix or vector.
}
\description{
Given a matrix with n rows (or a vector of length n), samples n rows with replacement from the original matrix.
}
\examples{
boot.samp(rnorm(10))
}
\keyword{Bootstrapping}
