% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{compare.tail.to.normal}
\alias{compare.tail.to.normal}
\title{Compare tails of a sample to normal expectation.}
\usage{
compare.tail.to.normal(x, k, mu, sigma)
}
\arguments{
\item{x}{A numeric vector}

\item{k}{The number of standard deviations beyond which is considered the "tail" (for the purpose of the computation)}

\item{mu}{The expectation of a normal distribution to which x will be compared}

\item{sigma}{The standard deviation of a normal distribution to which x will be compared}
}
\value{
A ratio of the proportion of data beyond k standard deviations from the expectation divided by the Normal-theory probability of observations beyond k standard deviations from the expectation.
}
\description{
Compares the proportion of observations in a vector beyond k standard deviations to the expectation under a normal distribution. The output is a ratio of the proportion of data beyond k standard deviations from the expectation divided by the Normal-theory probability of observations beyond k standard deviations from the expectation.
}
\examples{
compare.tail.to.normal(rnorm(10000, 0, 1), 3, 0, 1)
compare.tail.to.normal(rt(10000, 2), 3, 0, 1)
}
\keyword{central}
\keyword{extreme}
\keyword{limit}
\keyword{observations,}
\keyword{theorem}
