% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{laplace.samps}
\alias{laplace.samps}
\title{Generate a matrix of samples from a Laplace distribution.}
\usage{
laplace.samps(mu = 0, sigma = 1, n = 25, nsamps = 10000)
}
\arguments{
\item{mu}{The expectation of the Laplace distribution from which to draw samples.}

\item{sigma}{The standard deviation of the Laplace distribution from which to draw samples.}

\item{n}{The number of independent observations to include in each sample.}

\item{nsamps}{The number of samples to generate.}
}
\value{
A matrix of independent Laplace-distributed random numbers with nsamps rows and n columns.
}
\description{
Draws Laplace samples and formats them into a matrix, where each row contains a sample.
}
\examples{
laplace.samps(10, 1, 5, 8)
}
\keyword{Laplace}
\keyword{distribution}
\keyword{simulation,}
