% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{norm.samps}
\alias{norm.samps}
\title{Generate a matrix of samples from a normal distribution.}
\usage{
norm.samps(mu = 0, sigma = 1, n = 25, nsamps = 10000)
}
\arguments{
\item{mu}{The expectation of the normal distribution from which to draw samples.}

\item{sigma}{The standard deviation of the normal distribution from which to draw samples.}

\item{n}{The number of independent observations to include in each sample.}

\item{nsamps}{The number of samples to generate.}
}
\value{
A matrix of independent normally distributed random numbers with nsamps rows and n columns.
}
\description{
Draws normal samples and formats them into a matrix, where each row contains a sample.
}
\examples{
norm.samps(10, 1, 5, 8)
}
\keyword{distribution}
\keyword{normal}
\keyword{simulation,}
