% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{perm.samp}
\alias{perm.samp}
\title{Randomly permute the columns of a matrix independently}
\usage{
perm.samp(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix resulting from randomly permuting the columns of x independently.
}
\description{
Given a matrix, return a version with its columns randomly and independently permuted.
}
\examples{
perm.samp(matrix(1:25, nrow = 5))
}
\keyword{permutation}
