% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{post.conj.norm.norm}
\alias{post.conj.norm.norm}
\title{Compute normal conjugate posterior}
\usage{
post.conj.norm.norm(z, known.sd, prior.mean, prior.sd)
}
\arguments{
\item{z}{A numeric vector (assumed to be drawn from a normal distribution with known variance / standard deviation)}

\item{known.sd}{The known standard deviation of the data distribution.}

\item{prior.mean}{The expectation of the normal prior.}

\item{prior.sd}{The standard devation of the normal prior.}
}
\value{
A list containing the mean and standard deviation of the conjugate posterior distribution.
}
\description{
Given a sample of data from a normal distribution with known variance / standard deviation, and a normal prior for the expectation of the data, compute the parameters of the posterior for the expectation.
}
\examples{
post.conj.norm.norm(rnorm(100), 1, 2, 4)
}
\keyword{Conjugate}
\keyword{Normal,}
\keyword{posterior.}
\keyword{prior,}
