% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{power.sim.1sz}
\alias{power.sim.1sz}
\title{Estimate the power of a one-sample z test by simulation}
\usage{
power.sim.1sz(n, nsim, d, lev = 0.05)
}
\arguments{
\item{n}{The size of each simulated sample.}

\item{nsim}{The number of simulated samples to run.}

\item{d}{An effect size: the number of (known) standard deviations by which the expectation of the sampled data differs from a hypothesized expectation.}

\item{lev}{The significance level of the one-sample z-test.}
}
\value{
The proportion of p values less than the specified level (i.e. the estimated power).
}
\description{
Simulates nsims normal samples with expectations differing from a value specified by the null hypothesis by d (known) standard deviations. For each sample, the null hypothesis is tested by a one-sample z test, and the p value is compared with the specified significance level.
}
\examples{
powersim.1sz(n = 20, nsim = 10000, d = .5)
}
\keyword{power,}
\keyword{simulation,}
\keyword{test}
\keyword{z}
