% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{ps.1sz}
\alias{ps.1sz}
\title{Estimate the power of a one-sample z test by simulation}
\usage{
ps.1sz(d, n, level = 0.05, nsim = 10000)
}
\arguments{
\item{d}{An effect size: the number of (known) standard deviations by which the expectation of the sampled data differs from a hypothesized expectation.}

\item{n}{The size of each simulated sample.}

\item{level}{The significance level of the one-sample z-test.}

\item{nsim}{The number of simulated samples to run.}
}
\value{
The proportion of p values less than the specified level (i.e. the power).
}
\description{
Simulates nsims normal samples with expectations differing from a value specified by the null hypothesis by d (known) standard deviations. For each sample, the null hypothesis is tested by a one-sample z test, and the p value is compared with the specified significance level.
}
\examples{
ps.1sz(d = 0.5, n = 20)
}
\keyword{power,}
\keyword{simulation,}
\keyword{test}
\keyword{z}
