% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{reject.samp.norm}
\alias{reject.samp.norm}
\title{Perform rejection sampling for normally distributed data with a normal prior on the expectation.}
\usage{
reject.samp.norm(z, known.sd = 1, prior.mn = 0, prior.sd = 1,
  nsamps = 10000)
}
\arguments{
\item{z}{A numeric vector (assumed to be drawn from a normal distribution with known variance / standard deviation)}

\item{known.sd}{The known standard deviation of the data distribution.}

\item{prior.mn}{The expectation of the normal prior.}

\item{prior.sd}{The standard devation of the normal prior.}

\item{nsamps}{The number of observations to draw from the posterior by rejection sampling.}
}
\value{
A numeric vector of length nsamps containing simulated data from the posterior distribution.
}
\description{
Given a sample of data from a normal distribution with known variance / standard deviation, and a normal prior for the expectation of the data, draw samples from the posterior by rejection sampling.
}
\examples{
reject.samp.norm(rnorm(100), 1, 2, 4)
}
\keyword{Rejection}
\keyword{normal,}
\keyword{posterior.}
\keyword{sampling,}
