% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{rlaplace}
\alias{rlaplace}
\title{Sample from a Laplace (also called double-exponential) distribution}
\usage{
rlaplace(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{The number of independent observations to draw (i.e. sample size).}

\item{mean}{The expectation of the Laplace distribution.}

\item{sd}{The standard deviation of the Laplace distribution.}
}
\value{
A vector of Laplace-distributed random numbers of length n.
}
\description{
Simulates a sample of Laplace-distributed data.
}
\examples{
rlaplace(10, 0, 1)
}
\keyword{Laplace,}
\keyword{data,}
\keyword{double-exponential}
\keyword{simulated}
