% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{rnorm.contam}
\alias{rnorm.contam}
\title{Sample from a Normal distribution with outliers}
\usage{
rnorm.contam(n, mu = 0, sigma = 1, contam.p = 0.01, contam.mu = -5,
  contam.sigma = 0)
}
\arguments{
\item{n}{The number of independent observations per sample}

\item{mu}{The expectation of the target distribution.}

\item{sigma}{The standard deviation of the target distribution.}

\item{contam.p}{The probability that each observation is from the non-target distribution.}

\item{contam.mu}{The expectation of the non-target / "contamination" / outlier distribution.}

\item{contam.sigma}{The standard deviation of the outlier distribution.}
}
\value{
A vector of independent random numbers n entries.
}
\description{
Simulates a samples of normally distributed data, each "contaminated" with outliers from a different normal distribution according to a user-specified probability. Similar to rnorm.out, except that the order of arguments is different and the output is formatted as a vector.
}
\examples{
rnorm.mix(20, 0, 1, .2, 10, 1)
}
\keyword{data,}
\keyword{distribution,}
\keyword{normal}
\keyword{outliers}
\keyword{simulated}
