% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{rnorm.out}
\alias{rnorm.out}
\title{A matrix of samples from a Normal distribution with outliers}
\usage{
rnorm.out(nsamps, n, gamma, theta = 0, lambda = 10, targ.sd = 1,
  out.sd = 1)
}
\arguments{
\item{nsamps}{The number of samples to simulate.}

\item{n}{The number of independent observations per sample}

\item{gamma}{The probability that each observation is from the non-target distribution.}

\item{theta}{The expectation of the target distribution.}

\item{lambda}{The expectation of the non-target / "contamination" / outlier distribution.}

\item{targ.sd}{The standard deviation of the target distribution.}

\item{out.sd}{The standard deviation of the outlier distribution.}
}
\value{
A matrix of independent random numbers with nsamps rows and n columns.
}
\description{
Simulates samples of normally distributed data, each "contaminated" with outliers from a different normal distribution according to a user-specified probability.
}
\examples{
rnorm.out(5, 20, .2, 0, 10)
}
\keyword{data,}
\keyword{distribution,}
\keyword{normal}
\keyword{outliers}
\keyword{simulated}
