% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{rpareto}
\alias{rpareto}
\title{Sample from a Pareto distribution}
\usage{
rpareto(n, a = 0.5, b = 1)
}
\arguments{
\item{n}{The number of independent observations to draw (i.e. sample size).}

\item{a}{The shape parameter of the Pareto distribution.}

\item{b}{The scale parameter of the Pareto distribution.}
}
\value{
A vector of Pareto-distributed random numbers of length n.
}
\description{
Simulates a sample of Pareto-distributed data.
}
\examples{
rpareto(100, 0.5, 1)
}
\keyword{Pareto}
\keyword{data,}
\keyword{simulated}
