% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{sim.2var}
\alias{sim.2var}
\title{Simulate data from a a linear model with an unobserved confounder.}
\usage{
sim.2var(n, nsim, a, b1, b2 = 0, sigma.disturb = 1, correl = 0)
}
\arguments{
\item{n}{The number of independent pairs of observations to generate per simulation.}

\item{nsim}{The number of simulations to run.}

\item{a}{The intercept.}

\item{b1}{The slope for the observed independent variable.}

\item{b2}{The slope for the unobserved confounder.}

\item{sigma.disturb}{The standard deviation of the disturbances.}

\item{correl}{The correlation of (observed) x and (unobserved) z.}
}
\value{
A list of two matrices. The first matrix contains the x values from each simulation, with one simuluation in each row. The second matrix contains the y values in the same configuration.
}
\description{
Generates nsim samples of n independent pairs of observations using a linear model y = a + b1*x + b2*z + disturbance where z is unobserved. x and z are jointly normally distributed, and the disturbances are also normal.
}
\examples{
sim.2var(10, 5, a = 3, b1 = 1/2, b2 = 1/5, sigma.disturb = 1, correl = .5)
}
\keyword{Simulation,}
\keyword{bias.}
\keyword{confounding,}
\keyword{linear}
\keyword{omitted}
\keyword{regression,}
\keyword{simple}
\keyword{variable}
