% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{sim.Wald.B}
\alias{sim.Wald.B}
\title{Simulate simple linear regression data and apply a Wald test}
\usage{
sim.Wald.B(n, nsim, a, b, B0 = 0, sigma.disturb = 1, mu.x = 8,
  sigma.x = 2, rdisturb = rnorm, rx = rnorm, het.coef = 0,
  pfun = pnorm, ...)
}
\arguments{
\item{n}{The number of pairs of observations to simulate.}

\item{nsim}{The number of simulations to conduct.}

\item{a}{The intercept parameter of the linear regression model to be simulated.}

\item{b}{The slope parameter of the linear regression model to be simulated.}

\item{B0}{The value of the slope under the null hypothesis to be tested.}

\item{sigma.disturb}{The variance of the disturbances in the model y = a + b*x + disturbance.}

\item{mu.x}{The expectation of the x values.}

\item{sigma.x}{The variance of the x values.}

\item{rdisturb}{A function for drawing the random disturbances. rnorm() is the default, which makes the disturbances normally distributed, but you can use any function for random number generation with first argument the sample size, second argument the expectation, and third argument the standard deviation.}

\item{rx}{A function for drawing the random x values. rnorm() is the default, which makes x normally distributed, but you can use any function for random number generation with first argument the sample size, second argument the expectation, and third argument the standard deviation.}

\item{het.coef}{A number introducing some heteroscedasticity (i.e. non-constant variance) to the disturbances. If het.coef = 0 (the default), then the disturbances have constant variance. If it is positive, then the standard deviation of the disturbances increases with x; if negative, then the standard deviation of the disturbances decreases with x.}

\item{pfun}{A cumulative distribution function to which to compare the Wald statistic.}

\item{...}{Additional arguments to pfun}
}
\value{
A vector of p values of length nsims, one from each Wald test.
}
\description{
Simulates data from a linear model using the sim.lm() function, each time applying a Wald test test the null hypothesis that the slope = B0.
}
\examples{
sim.Wald.B(10, 100, 3, .1)
}
\keyword{Wald}
\keyword{data,}
\keyword{linear}
\keyword{regression,}
\keyword{simple}
\keyword{simulated}
