% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{wrap.bm}
\alias{wrap.bm}
\title{Simulate a normal sample and compute a bootstrap distribution}
\usage{
wrap.bm(n, B, mu = 0, sigma = 1, FUN = mean, ...)
}
\arguments{
\item{n}{The size of the normal sample to draw.}

\item{B}{The number of bootstrap samples to draw from the normal sample.}

\item{mu}{The expectation of the normal distribution from which the data are drawn.}

\item{sigma}{The standard deviation of the normal distribution from which the data are drawn.}

\item{FUN}{A function that computes the statistic of interest. FUN must have a vector of data as its first argument and must return a numeric vector of length 1 (i.e. a scalar).}

\item{...}{Additional arguments passed to FUN.}
}
\value{
A list containing the mean and standard deviation of the bootstrap distribution.
}
\description{
Wraps boot.dist.1d, generating a normal sample and computing a bootstrap distribution for a statistic of interest. Plots a histogram of the bootstrap distribution and returns a list containing the mean and standard deviation of the bootstrap distribution.
}
\examples{
wrap.bm(50, 1000, FUN = median)
}
\keyword{Bootstrap}
\keyword{Bootstrapping,}
\keyword{distribution}
