% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{many.outcome.sim}
\alias{many.outcome.sim}
\title{Simulate a multiple-testing scenario}
\usage{
many.outcome.sim(n, n.meas, correl, n.sims)
}
\arguments{
\item{n}{The number of people measured in each group per simulation.}

\item{n.meas}{The number of outcome measurements taken per person.}

\item{correl}{The correlation of the (normally distributed) outcome measurements---all pairs of measurements have the same correlation.}

\item{n.sims}{The number of simulated samples to run.}
}
\value{
A matrix of p values with n.meas columns and n.sims rows. Each row corresponds to a different sample, and the p values from the respective outcome measures are in distinct columns.
}
\description{
Simulates type I error inflation from multiple testing for a test comparing two group means. The groups are compared on n.meas outcome measurements, each mutually correlated at a user-specified level. A matrix of p values is returned with n.meas columns and n.sims rows.
}
\examples{
many.outcome.sim(100, 3, .5, 10)
}
\keyword{data,}
\keyword{hacking}
\keyword{multiple}
\keyword{p}
\keyword{simulated}
\keyword{testing,}
