% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{sim.perm.B}
\alias{sim.perm.B}
\title{Simulate simple linear regression data and apply a permutation test}
\usage{
sim.perm.B(n, nsim, a, b, nperm = 500, sigma.disturb = 1, mu.x = 8,
  sigma.x = 4, rdisturb = rnorm, rx = rnorm, het.coef = 0)
}
\arguments{
\item{n}{The number of pairs of observations to simulate.}

\item{nsim}{The number of simulations to conduct.}

\item{a}{The intercept parameter of the linear regression model to be simulated.}

\item{b}{The slope parameter of the linear regression model to be simulated.}

\item{nperm}{The number of permutations to use per simulation to conduct the permutation test.}

\item{sigma.disturb}{The variance of the disturbances in the model y = a + b*x + disturbance.}

\item{mu.x}{The expectation of the x values.}

\item{sigma.x}{The variance of the x values.}

\item{rdisturb}{A function for drawing the random disturbances. rnorm() is the default, which makes the disturbances normally distributed, but you can use any function for random number generation with first argument the sample size, second argument the expectation, and third argument the standard deviation.}

\item{rx}{A function for drawing the random x values. rnorm() is the default, which makes x normally distributed, but you can use any function for random number generation with first argument the sample size, second argument the expectation, and third argument the standard deviation.}

\item{het.coef}{A number introducing some heteroscedasticity (i.e. non-constant variance) to the disturbances. If het.coef = 0 (the default), then the disturbances have constant variance. If it is positive, then the standard deviation of the disturbances increases with x; if negative, then the standard deviation of the disturbances decreases with x.}
}
\value{
A vector of p values of length nsims, one from each permutation test of the hypothesis that the slope is zero.
}
\description{
Simulates data from a linear model using the sim.lm() function, each time applying a permutation test to test the null hypothesis that the slope is zero.
}
\examples{
ps <- sim.perm.B(10, 200, a = 3, b = .1, nperm = 200)
mean(ps < .05) #power at .05 significance level
}
\keyword{data,}
\keyword{linear}
\keyword{permutation}
\keyword{regression,}
\keyword{simple}
\keyword{simulated}
\keyword{test}
