% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{sim.Wald.B}
\alias{sim.Wald.B}
\title{Simulate simple linear regression data and apply a Wald test}
\usage{
sim.Wald.B(a, b, B0 = 0, n.sim = 1000, var.eps = 1, n = 50,
  mu.x = 8, var.x = 4, rdist = rnorm, rx = rnorm, pfun = pnorm,
  ...)
}
\arguments{
\item{a}{The intercept parameter of the linear regression model to be simulated.}

\item{b}{The slope parameter of the linear regression model to be simulated.}

\item{B0}{The value of the slope under the null hypothesis to be tested.}

\item{n.sim}{The number of simulations to conduct.}

\item{var.eps}{The variance of the disturbances in the model y = a + b*x + disturbance.}

\item{n}{The numer of independent pairs of observations to simulate per sample.}

\item{mu.x}{The expectation of the x values.}

\item{var.x}{The variance of the x values.}

\item{rdist}{A function for drawing the random disturbances. rnorm() is the default, which makes the disturbances normally distributed, but you can use any function for random number generation with first argument the sample size, second argument the expectation, and third argument the standard deviation.}

\item{rx}{A function for drawing the random x values. rnorm() is the default, which makes x normally distributed, but you can use any function for random number generation with first argument the sample size, second argument the expectation, and third argument the standard deviation.}

\item{pfun}{A cumulative distribution function to which to compare the Wald statistic.}

\item{...}{Additional arguments to pfun}
}
\value{
A vector of p values of length nsims, one from each Wald test.
}
\description{
Simulates data from a linear model using the sim.lm() function, each time applying a Wald test test the null hypothesis that the slope = B0.
}
\examples{
sim.Wald.B(3, .1, n.sim = 20)
}
\keyword{Wald}
\keyword{data,}
\keyword{linear}
\keyword{regression,}
\keyword{simple}
\keyword{simulated}
